pyproject.toml
# Poetry section
[tool.poetry]
name = "sit-backend"
version = "0.1.0"
description = ""
authors = ["Sven <sven.hoyer@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.1"
djangorestframework = "^3.14.0"
channels = {extras = ["daphne"], version = "^4.0.0"}
djangorestframework-simplejwt = "^5.3.1"
djangochannelsrestframework = "^1.2.0"
django-channels-jwt-auth-middleware = "^1.0.0"
django-filter = "^23.5"
django-cors-headers = "^4.3.1"
numpy = "^1.26.3"
channels-redis = "^4.2.0"
pyswarms = "^1.3.0"


[tool.poetry.group.test.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
pytest-django = "^4.7.0"


[tool.poetry.group.dev.dependencies]
black = "^23.12.1"
pylint = "^3.0.3"
mypy = "^1.8.0"
django-stubs = "^4.2.7"
pylint-django = "^2.5.5"
ipykernel = "^6.29.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


# Style and Linting Configuration	
#black section
[tool.black]
line-length=79
force-exclude = '''
/(
  | ^.*\b(migrations)\b.*$
)/

'''
#isort section
[tool.isort]
profile = "black"

#Pylint section
[tool.pylint]

[tool.pylint.main]
disable = ['I1101', 'C0114', 'C0115', 'C0116', 'W1203', 'W0621', 'R0903']
ignore="manage.py"
ignore-paths='.*/migrations/*, ./config/*, .*/tests/*'
max-attributes=20
max-locals=20

[tool.pylint.logging]
logging-format-style = "new"

# MyPy section
[tool.mypy]
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]

[tool.django-stubs]
django_settings_module = "config.settings"

#Pytest section
[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = 'config.settings.test'
python_files = ['tests.py', 'test_*.py']
filterwarnings = 'ignore::DeprecationWarning'
